src/requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.5
durationpy==0.10
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
kubernetes==35.0.0
oauthlib==3.3.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
smmap==5.0.3
urllib3==2.6.3
websocket-client==1.9.0
